broke out code so that I can start building different commands starting with npm-deep
