+++ bin/npm-deep.js
@@ -1,15 +1,12 @@
 #!/usr/bin/env node
 
-
-const npmrec = require("../index");
+const npmExec = require("../src/npm-exec");
+const utils = require("belty");
 const argv = require("subarg")(process.argv.slice(2));
 const command = argv._.shift();
 const files = argv._.slice(0);
 
-// Remove unwanted key.
-delete argv._;
-
-npmrec(command, Object.assign({ files: files }, argv))
+npmExec(command, Object.assign({ files: files }, utils.omit(argv, ["_"])))
   .then(() => console.log("All set"))
   .catch((ex) => {
     ex.errors.forEach(error => {
